Refactors form components
